+++ langs/fire/FireConstants.js
@@ -1,0 +1,5 @@
+const FireConstants = {}
+
+FireConstants.fileExtension = ".fire"
+
+module.exports = FireConstants