+++ built/jtree.browser.js
@@ -430,9 +430,6 @@ class ImmutableNode extends AbstractNode {
     hasDuplicateKeywords() {
         return this.length ? new Set(this.getKeywords()).size !== this.length : false;
     }
-    _isAllMaps() {
-        return !this.hasDuplicateKeywords() && this.every(node => !node.hasDuplicateKeywords());
-    }
     isEmpty() {
         return !this.length && !this.getContent();
     }
@@ -1787,19 +1784,30 @@ class TreeNode extends ImmutableNode {
         this.forEach(node => node.delete(keyword));
         return this;
     }
+    _getNonMaps() {
+        const results = this.getTopDownArray().filter(node => node.hasDuplicateKeywords());
+        if (this.hasDuplicateKeywords())
+            results.unshift(this);
+        return results;
+    }
     // todo: add more testing.
     // todo: solve issue with where extend should overwrite or append
+    // todo: should take a grammar? to decide whether to overwrite or append.
     extend(nodeOrStr) {
         if (!(nodeOrStr instanceof TreeNode))
             nodeOrStr = new TreeNode(nodeOrStr);
-        if (!nodeOrStr._isAllMaps())
-            throw new Error(`Currently extend only works with maps but the given tree contains duplicate keys at some level.`);
-        nodeOrStr.forEach(node => {
-            const path = node.getKeyword();
-            const content = node.getContent();
-            const targetNode = this.touchNode(path).setContent(content);
-            if (node.length)
-                targetNode.extend(node.childrenToString());
+        const usedKeywords = new Set();
+        nodeOrStr.forEach(sourceNode => {
+            const keyword = sourceNode.getKeyword();
+            let targetNode;
+            if (usedKeywords.has(keyword))
+                targetNode = this.appendLine(sourceNode.getLine());
+            else {
+                targetNode = this.touchNode(keyword).setContent(sourceNode.getContent());
+                usedKeywords.add(keyword);
+            }
+            if (sourceNode.length)
+                targetNode.extend(sourceNode.childrenToString());
         });
         return this;
     }
@@ -4060,4 +4068,4 @@ jtree.AnyNode = GrammarBackedAnyNode;
 jtree.GrammarProgram = GrammarProgram;
 jtree.UnknownGrammarProgram = UnknownGrammarProgram;
 jtree.TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode;
-jtree.getVersion = () => "22.0.0";
+jtree.getVersion = () => "22.1.0";